# utilitatsfirma-api-interna/openapi/openapi_comanda_v1.json
{
  "openapi" : "3.0.1",
  "servers" : [ {
    "url" : "/utilitatsfirmaapi/interna",
    "description" : "Servidor local (sense protocol ni host) per a desplegaments en entorns on el context de l'aplicació és /utilitatsfirmaapi/interna)"
  }, {
    "url" : "http://localhost:8080/utilitatsfirmaapi/interna"
  }, {
    "url" : "https://dev.caib.es/utilitatsfirmaapi/interna"
  }, {
    "url" : "https://proves.caib.es/utilitatsfirmaapi/interna"
  }, {
    "url" : "https://se.caib.es/utilitatsfirmaapi/interna"
  }, {
    "url" : "https://www.caib.es/utilitatsfirmaapi/interna"
  } ],
  "paths" : {
...
    "/secure/logs/v1/{nomFitxer}/directe" : {
      "get" : {
        "operationId" : "descarregarFitxerDirecte",
        "parameters" : [ {
          "name" : "nomFitxer",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/octet-stream" : { }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/logs/v1/{nomFitxer}" : {
      "get" : {
        "operationId" : "getFitxerByNom",
        "parameters" : [ {
          "name" : "nomFitxer",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FitxerContingut"
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/logs/v1/{nomFitxer}/linies/{nLinies}" : {
      "get" : {
        "operationId" : "llegitUltimesLinies",
        "parameters" : [ {
          "name" : "nomFitxer",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "nLinies",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/logs/v1" : {
      "get" : {
        "operationId" : "llistarFitxers",
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FitxerInfo"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "FitxerContingut" : {
        "type" : "object",
        "properties" : {
          "nom" : {
            "type" : "string"
          },
          "mida" : {
            "type" : "integer",
            "format" : "int64"
          },
          "mimeType" : {
            "type" : "string"
          },
          "dataCreacio" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dataModificacio" : {
            "type" : "string",
            "format" : "date-time"
          },
          "contingut" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "byte"
            }
          }
        }
      },
      "FitxerInfo" : {
        "type" : "object",
        "properties" : {
          "nom" : {
            "type" : "string"
          },
          "mida" : {
            "type" : "integer",
            "format" : "int64"
          },
          "mimeType" : {
            "type" : "string"
          },
          "dataCreacio" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dataModificacio" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      }
    }
  }
}